refactor: DTO 클래스 구조 및 관련 로직 개선
- DTO 클래스들을 요청 및 응답별로 파일로 분리
- toResponse, toListResponse 확장 함수로 변경
