--- gs/src/main/kotlin/com/pluxity/patrol/dto/ScenarioExecutionFailRequest.kt
@@ -1,5 +1,0 @@
-package com.pluxity.patrol.dto
-
-data class ScenarioExecutionFailRequest(
-    val errorMessage: String? = null,
-)
