+++ gs/src/main/kotlin/com/pluxity/patrol/dto/SceneResponse.kt
@@ -4,6 +4,7 @@ import com.pluxity.feature.entity.Spatial
 import com.pluxity.patrol.constant.DeviceAction
 import com.pluxity.patrol.constant.DeviceType
 import com.pluxity.patrol.entity.Scene
+import com.pluxity.patrol.entity.SceneDeviceAction
 
 data class SceneResponse(
     val id: Long,
@@ -28,34 +29,6 @@ data class SceneResponse(
         val actionParam: String?,
         val executionOrder: Int?,
     )
-
-    companion object {
-        fun from(scene: Scene): SceneResponse =
-            SceneResponse(
-                id = scene.requiredId,
-                name = scene.name,
-                description = scene.description,
-                duration = scene.duration,
-                position = scene.position,
-                rotation = scene.rotation,
-                facility =
-                    FacilitySummary(
-                        id = scene.facility.requiredId,
-                        name = scene.facility.name,
-                    ),
-                sceneDeviceActions =
-                    scene.sceneDeviceActions.map { action ->
-                        SceneDeviceActionResponse(
-                            sceneDeviceActionId = action.requiredId,
-                            deviceId = action.deviceId,
-                            deviceType = action.deviceType,
-                            deviceAction = action.deviceAction,
-                            actionParam = action.actionParam,
-                            executionOrder = action.executionOrder,
-                        )
-                    },
-            )
-    }
 }
 
 data class SceneListResponse(
@@ -63,14 +36,34 @@ data class SceneListResponse(
     val name: String,
     val description: String?,
     val duration: Int?,
-) {
-    companion object {
-        fun from(scene: Scene): SceneListResponse =
-            SceneListResponse(
-                id = scene.requiredId,
-                name = scene.name,
-                description = scene.description,
-                duration = scene.duration,
-            )
-    }
-}
+)
+
+fun Scene.toResponse() =
+    SceneResponse(
+        id = requiredId,
+        name = name,
+        description = description,
+        duration = duration,
+        position = position,
+        rotation = rotation,
+        facility = SceneResponse.FacilitySummary(facility.requiredId, facility.name),
+        sceneDeviceActions = sceneDeviceActions.map { it.toResponse() },
+    )
+
+fun SceneDeviceAction.toResponse() =
+    SceneResponse.SceneDeviceActionResponse(
+        sceneDeviceActionId = requiredId,
+        deviceId = deviceId,
+        deviceType = deviceType,
+        deviceAction = deviceAction,
+        actionParam = actionParam,
+        executionOrder = executionOrder,
+    )
+
+fun Scene.toListResponse() =
+    SceneListResponse(
+        id = requiredId,
+        name = name,
+        description = description,
+        duration = duration,
+    )
