--- gs/src/main/kotlin/com/pluxity/patrol/dto/TriggerTargetRequest.kt
@@ -1,8 +1,0 @@
-package com.pluxity.patrol.dto
-
-import com.pluxity.patrol.constant.TriggerTargetType
-
-data class TriggerTargetRequest(
-    val targetType: TriggerTargetType = TriggerTargetType.USER,
-    val targetId: String,
-)
